// apps/web/src/api/__generated__/collection/collection.ts
export type getCollectionPostsResponse200 = {
  data: GetCollectionPostsResponse;
  status: 200;
};

export type getCollectionPostsResponseSuccess =
  getCollectionPostsResponse200 & {
    headers: Headers;
  };
export type getCollectionPostsResponse = getCollectionPostsResponseSuccess;